pyproject.toml
[tool.poetry]
name = "iatt"
version = "0.1.0"
description = ""
authors = ["Phil Leggetter <phil@leggetter.co.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
pymongo = "^4.8.0"
replicate = "^0.32.1"
gunicorn = "^23.0.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
